golangci-lint: improve handling of generated files
Big autogenerated files take too much time and memory to load,
even if we skip them with -skip-dirs or -skip-dirs; or mark them as generated; or use nolint annotations.

So here we define a tag and use it in the autogenerated files.

Also updated Stratum generated file manually, since regenerating causes changes that are not relevant to this commit.

Future models' generated.go files however will get updated automatically by the updated ModelGenerator.sh.
Deliberately not changing the other generated.go files that are used today in UTs i.e.
 TestDevice-1.0.0 TestDevice-2.0.0 Devicesim-1.0.0

To clearly see effect of this change use -v e.g.
 golangci-lint run --timeout 30m -v
For big generated model files (which are expected to be inroduced to µonos in the near future), the savings are significant.